+++ packages/xrpl/src/models/transactions/transaction.ts
@@ -57,6 +57,7 @@ import {
   validateNFTokenCreateOffer,
 } from './NFTokenCreateOffer'
 import { NFTokenMint, validateNFTokenMint } from './NFTokenMint'
+import { NFTokenModify, validateNFTokenModify } from './NFTokenModify'
 import { OfferCancel, validateOfferCancel } from './offerCancel'
 import { OfferCreate, validateOfferCreate } from './offerCreate'
 import { OracleDelete, validateOracleDelete } from './oracleDelete'
@@ -151,6 +152,7 @@ export type SubmittableTransaction =
   | NFTokenCancelOffer
   | NFTokenCreateOffer
   | NFTokenMint
+  | NFTokenModify
   | OfferCancel
   | OfferCreate
   | OracleDelete
@@ -385,6 +387,10 @@ export function validate(transaction: Record<string, unknown>): void {
 
     case 'NFTokenMint':
       validateNFTokenMint(tx)
+      break
+
+    case 'NFTokenModify':
+      validateNFTokenModify(tx)
       break
 
     case 'OfferCancel':