+++ packages/ripple-binary-codec/HISTORY.md
@@ -11,6 +11,7 @@
 
 ### Added
 * Support for the Price Oracles amendment (XLS-47).
+* Add `NFTokenModify` transaction and add `tfMutable` flag in `NFTokenMint`
 
 ### Fixed
 * Better error handling/error messages for serialization/deserialization errors.